organize location of files and folders
