+++ frontend/components/category/product/CategoryProduct.scss
@@ -1,0 +1,11 @@
+.cat.product {
+  p {
+    font-size: 16px;
+  }
+    .img {
+      width: 100%;
+      max-width: 142px;
+      max-height: 112px;
+      object-fit: cover;    
+    }    
+}
