--- frontend/modules/category/CategoryProduct.scss
@@ -1,8 +1,0 @@
-.product {
-  .img {
-    width: 100%;
-    max-width: 142px;
-    max-height: 112px;
-    object-fit: cover;    
-  }
-}
